+++ src/test/java/com/project/kodesalon/model/member/controller/MemberControllerTest.java
@@ -2,9 +2,8 @@ package com.project.kodesalon.model.member.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.project.kodesalon.common.GlobalExceptionHandler;
-import com.project.kodesalon.model.member.controller.dto.ChangePasswordRequest;
 import com.project.kodesalon.model.member.service.MemberService;
-import com.project.kodesalon.model.member.service.dto.ChangePasswordRequestDto;
+import com.project.kodesalon.model.member.service.dto.ChangePasswordRequest;
 import com.project.kodesalon.model.member.service.dto.ChangePasswordResponseDto;
 import com.project.kodesalon.model.member.service.dto.CreateMemberRequest;
 import com.project.kodesalon.model.member.service.dto.LoginRequest;
@@ -311,7 +310,7 @@ public class MemberControllerTest {
     @Test
     @DisplayName("비밀번호 변경시, 변경하려는 비밀번호, 회원 식별 번호를 전달받아 비밀번호를 변경하고 200 상태 + 성공 메세지를 반환합니다.")
     public void changePassword() throws Exception {
-        given(memberService.changePassword(anyLong(), any(ChangePasswordRequestDto.class)))
+        given(memberService.changePassword(anyLong(), any(ChangePasswordRequest.class)))
                 .willReturn(new ChangePasswordResponseDto("비밀번호 변경 성공하였습니다."));
 
         this.mockMvc.perform(put("/api/v1/members/{memberId}", 1L)
@@ -353,7 +352,7 @@ public class MemberControllerTest {
     void failed_change_password_with_member_id_not_exist() throws Exception {
         ChangePasswordRequest changePasswordRequest = new ChangePasswordRequest("Password123!!");
 
-        given(memberService.changePassword(anyLong(), any(ChangePasswordRequestDto.class)))
+        given(memberService.changePassword(anyLong(), any(ChangePasswordRequest.class)))
                 .willThrow(new NoSuchElementException("찾으려는 회원이 없습니다"));
 
         this.mockMvc.perform(put("/api/v1/members/{memberId}", 1L)
