+++ src/test/java/com/project/kodesalon/model/member/service/dto/CreateMemberRequestTest.java
@@ -179,4 +179,41 @@ class CreateMemberRequestTest {
                 .extracting(ConstraintViolation::getMessage)
                 .contains("null이 아닌 휴대폰 번호를 입력해주세요.");
     }
+
+    static class ChangePasswordRequestTest {
+        private final ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
+        private final Validator validator = factory.getValidator();
+
+        @Test
+        @DisplayName("변경하려는 비밀번호를 반환한다.")
+        public void getter() {
+            ChangePasswordRequest changePasswordRequest = new ChangePasswordRequest("ChangePassword1!");
+            then(changePasswordRequest.getPassword()).isEqualTo("ChangePassword1!");
+        }
+
+        @ParameterizedTest
+        @ValueSource(strings = {"!pass12", "!!Password1234567", "Password12",
+                "!!Password", "!!password12", "!!PASSWORD12", "!비밀!pass1234"})
+        @DisplayName("Password가 유효하지 않으면 예외를 발생시킨다.")
+        void create_throws_exception_with_invalid_password(String invalidPassword) {
+            ChangePasswordRequest changePasswordRequestDto = new ChangePasswordRequest(invalidPassword);
+            Set<ConstraintViolation<ChangePasswordRequest>> constraintViolations = validator.validate(changePasswordRequestDto);
+
+            then(constraintViolations)
+                    .extracting(ConstraintViolation::getMessage)
+                    .contains("비밀번호는 영어 소문자, 대문자, 숫자, 특수문자를 포함한 8자리이상 16자리 이하여야 합니다.");
+        }
+
+        @ParameterizedTest
+        @NullSource
+        @DisplayName("Password가 Null이면 예외를 발생시킨다.")
+        void create_throws_exception_with_null_password(String nullPassword) {
+            ChangePasswordRequest changePasswordRequest = new ChangePasswordRequest(nullPassword);
+            Set<ConstraintViolation<ChangePasswordRequest>> constraintViolations = validator.validate(changePasswordRequest);
+
+            then(constraintViolations)
+                    .extracting(ConstraintViolation::getMessage)
+                    .contains("null이 아닌 8자리 이상의 비밀번호를 입력해주세요.");
+        }
+    }
 }
